Finishing Up Movie Details Activity

+++ app/src/main/java/com/example/themoviemvvm/data/value_object/MovieDetails.kt
@@ -14,7 +14,7 @@ data class MovieDetails(
     val posterPath: String,
     @SerializedName("release_date")
     val releaseDate: String,
-    val revenue: Int,
+    val revenue: Long, // In case of Connection Problem
     val runtime: Int,
     val tagline: String,
     val status: String,
